requirements.txt
# Kobe Trading System Dependencies
# Using minimum version specifiers for compatibility

# Core Data Processing
pandas>=2.0
numpy>=1.24
pyarrow>=14.0

# HTTP & API
requests>=2.31
aiohttp>=3.9
websockets>=12.0

# Broker Integration
alpaca-py>=0.13

# Configuration
pyyaml>=6.0
python-dotenv>=1.0

# Free data fallbacks (used by multi_source provider)
yfinance>=0.2.28
pandas-datareader>=0.10

# Technical Analysis
ta>=0.11.0
# pandas-ta requires Python 3.12+, install manually if needed:
# pip install pandas-ta  # 150+ technical indicators

# ML & AI Features
stumpy>=1.12  # Matrix profile for anomaly detection
scikit-learn>=1.3  # ML algorithms and preprocessing
scipy>=1.11  # Scientific computing
joblib>=1.3  # model serialization
vaderSentiment>=3.3  # sentiment analysis (VADER)
anthropic>=0.21  # For Claude API integration

# Visualization (for reports)
matplotlib>=3.7
plotly>=5.18

# Testing
pytest>=7.4
pytest-cov>=4.1
pytest-asyncio>=0.21

# Type Checking
mypy>=1.7

# Code Quality
black>=23.0
flake8>=6.1
isort>=5.12

# Scheduling
schedule>=1.2

# Date/Time
python-dateutil>=2.8
pytz>=2023.3

# HTTP Server (for health endpoints)
flask>=3.0

# Dashboard (Mission Control)
fastapi>=0.104
uvicorn>=0.24

# Alerts & Notifications
python-telegram-bot>=20.0

# Bayesian Optimization
optuna>=3.4

# Market Regime Detection (HMM)
hmmlearn>=0.3

# Ensemble ML (optional - for advanced ML features)
# Install these manually for full ensemble functionality:
# pip install xgboost>=2.0
# pip install lightgbm>=4.0
